dexter/dexter-agent/requirements.txt
# Core Dependencies
openai>=1.0.0

# Data Processing & Analysis
numpy>=1.26.0
pandas>=2.2.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
typing-extensions>=4.9.0

# Optional: Enhanced Data Visualization
tabulate>=0.9.0

# Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
mypy>=1.8.0
